FS-1581: Move location of transformation functionality
