+++ api/routes/assessment_routes.py
@@ -2,9 +2,7 @@
 from typing import Dict
 from typing import List
 
-from config.mappings.assessment_mapping_fund_round import (
-    transform_to_assessor_task_list_metadata,
-)
+from api.routes._helpers import transform_to_assessor_task_list_metadata
 from db.queries.assessment_records import get_metadata_for_fund_round_id
 from db.queries.assessment_records.queries import find_assessor_task_list_state
